Functional direct login
This branch gives the capability to directly log in from an external website. The user provides the pair username,passhash as an html query to /direct/. The pashhash is simply the output of control.sha1digest(username). At the moment this is insecure as anybody could go to the login page and enter username: username and password: username.